+++ frontend/admin/src/pages/userList/UserList.jsx
@@ -2,23 +2,22 @@ import "./userList.css";
 import { DataGrid } from "@mui/x-data-grid";
 import { DeleteOutline } from "@mui/icons-material";
 import { Link } from "react-router-dom";
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { deleteUser, getProducts, getUsers } from "../../redux/apiCalls";
+import { deleteUser, getUsers } from "../../redux/apiCalls";
 
 export default function UserList() {
   const dispatch = useDispatch();
   const users = useSelector((state) => state.user.users);
 
-  console.log("before useeffet");
-
   useEffect(() => {
-    console.log("effect");
     getUsers(dispatch);
   }, [dispatch]);
 
   const handleDelete = (id) => {
-    deleteUser(id, dispatch);
+    if (window.confirm("Are You Sure To Delete The User?")) {
+      deleteUser(id, dispatch);
+    }
   };
 
   const columns = [
@@ -85,6 +84,7 @@ export default function UserList() {
         columns={columns}
         getRowId={(row) => row._id}
         pageSize={8}
+        rowsPerPageOptions={[8]}
         checkboxSelection
       />
     </div>
